% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_molecular_weight.R
\name{calculate_molecular_weight}
\alias{calculate_molecular_weight}
\title{Calculate Molecular Weight of a Protein}
\usage{
calculate_molecular_weight(protein_sequence)
}
\arguments{
\item{protein_sequence}{A character string representing the protein sequence.}
}
\value{
A numeric value representing the molecular weight of the protein.
}
\description{
This function calculates the molecular weight of a protein sequence based on the
molecular weights of individual amino acids.
}
