% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_amino_seq.R
\name{protein_amino_seq}
\alias{protein_amino_seq}
\title{Determine Amino Acid Sequence in a Protein}
\usage{
protein_amino_seq(protein_sequence)
}
\arguments{
\item{protein_sequence}{A character string representing the protein sequence.}
}
\value{
A character string representing the amino acid sequence.
}
\description{
This function takes a protein sequence as input and returns the corresponding amino acid sequence.
}
